Special Chars| Re-factored automata into "Substitution DFAs".
A Substitution DFA (SDFA) carries a substitution table (buffer) and
the transitions may optionally contain an offset of a substitution.
An accepting state is one which is reached by an accepting transition.
An accepting transition is one which carries a non-negative offset into
the substitution table ("result").

After reading in the substitutions, the substitution table is "packed"
into a continuous buffer, so that it can be easily dumped and used from
a mmapped binary cache file.